Prepare for releasing docs on the website
Moved documentation to /docs/.
Replaced partial 2.6.x documentation with complete 2.7.x release.
Added a docs/index.md file for the docs directory to act as an index to the various versions provided on the site.
Add tzinfo gem to Gemfile so that jekyll can be run locally on Windows boxes.